# apps/runner/pkg/api/docs/swagger.yaml
definitions:
...
  ErrorResponse:
    description: Error response
    properties:
      code:
        example: BAD_REQUEST
        type: string
      message:
        example: Bad request
        type: string
      method:
        example: GET
        type: string
      path:
        example: /api/resource
        type: string
      statusCode:
        example: 400
        type: integer
      timestamp:
        example: '2023-01-01T12:00:00Z'
        type: string
    required:
      - message
      - path
      - statusCode
      - timestamp
    type: object
...
  generationstop.Fence:
    properties:
      workspaceExecutionManifestRef:
        type: string
    required:
      - workspaceExecutionManifestRef
    type: object
...
  workingcopy.CaptureCapabilities:
    properties:
      authority:
        $ref: '#/definitions/workingcopy.CaptureAuthority'
      fileSnapshot:
        $ref: '#/definitions/workingcopy.FileSnapshotCapability'
      stoppedWorkingTreeInventory:
        $ref: '#/definitions/workingcopy.WorkingTreeInventoryCapability'
    required:
      - authority
      - stoppedWorkingTreeInventory
    type: object
  workingcopy.CaptureCapabilitiesRequest:
    properties:
      authority:
        $ref: '#/definitions/workingcopy.CaptureAuthority'
      fence:
        $ref: '#/definitions/generationstop.Fence'
      owner:
        $ref: '#/definitions/workingcopy.CaptureOwner'
      source:
        $ref: '#/definitions/workingcopy.SourceAddress'
    required:
      - authority
      - fence
      - owner
      - source
    type: object
...
  workingcopy.CaptureIdentity:
    properties:
      authority:
        $ref: '#/definitions/workingcopy.CaptureAuthority'
      fileSnapshot:
        $ref: '#/definitions/workingcopy.FileSnapshotSource'
      owner:
        $ref: '#/definitions/workingcopy.CaptureOwner'
      providerName:
        type: string
      providerResourceId:
        type: string
      requestFingerprint:
        type: string
      selector:
        $ref: '#/definitions/workingcopy.CaptureSelector'
      source:
        $ref: '#/definitions/workingcopy.SourceAddress'
      stopAuthority:
        $ref: '#/definitions/generationstop.StopAuthority'
    required:
      - authority
      - owner
      - providerName
      - providerResourceId
      - requestFingerprint
      - selector
      - source
    type: object
...
info:
  contact: {}
  description: Daytona Runner API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Daytona Runner API
  version: v0.0.0-dev
paths:
...
  /sandboxes/{sandboxId}/working-copy-captures/capabilities:
    post:
      consumes:
        - application/json
      description: Read-only discovery checked against the Runner's measured capture
        component and physical source generation. Measurement does not replace image
        publication or conformance qualification.
      operationId: WorkingCopyCaptureCapabilities
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Current capture authority
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.CaptureCapabilitiesRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.CaptureCapabilities'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Discover the assigned Runner capture surface before stopping a generation
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/delete:
    post:
      operationId: DeleteWorkingCopyCapture
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Exact capture identity
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/workingcopy.CaptureIdentity'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.CaptureDeleteReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete an exact private working-copy capture
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/exists:
    post:
      operationId: WorkingCopyCaptureExists
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Exact capture identity
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/workingcopy.CaptureIdentity'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.CaptureExistsResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Check an exact private working-copy capture identity
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/observe:
    post:
      operationId: ObserveWorkingCopyCapture
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Exact capture binding
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/workingcopy.CaptureBinding'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.CaptureObservation'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Observe an exact working-copy capture
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/read:
    post:
      operationId: ReadWorkingCopyCapture
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Exact capture identity and read bounds
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/workingcopy.CaptureReadRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.CaptureReadResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Read an exact immutable working-copy capture
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: CaptureSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Capture an immutable native sandbox file once
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files/delete:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: DeleteSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileDeleteRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileDeleteReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Retire a native file operation and release its private bytes
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files/observe:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: ObserveSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileObserveRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileObservation'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Observe a native sandbox file operation without source effects
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/sandbox-files/read:
    post:
      consumes:
        - application/json
      description: Provider organization authority is supplied by the authenticated
        API, not a Product Run. File capture admits only work/outputs regular files
        and requires native descriptor cloning. Component measurement is not runtime
        qualification.
      operationId: ReadSandboxFile
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Native file capture operation
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileReadRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.SandboxFileReadResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Read a bounded immutable native file capture
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/stopped-directory-roster:
    post:
      description: Stream and validate a content-digested regular-file/directory roster
        through Docker host authority; archive link entries, special files, escapes,
        and bound overflows fail closed. Accepted regular-file entries are independent
        exact bytes; inode/link identity is not preserved.
      operationId: StoppedWorkingCopyDirectoryRoster
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Exact anchor, directory selector, and bounds
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/workingcopy.StoppedDirectoryRosterRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.StoppedDirectoryRosterReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List one bounded directory from an exact stopped sandbox generation
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/stopped-working-tree-inventories:
    post:
      consumes:
        - application/json
      operationId: PrepareWorkingTreeInventory
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Exact stopped working-tree inventory authority
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.WorkingTreeInventoryRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.WorkingTreeInventoryReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Prepare immutable pages of an exact stopped working tree
      tags:
        - sandbox
  /sandboxes/{sandboxId}/working-copy-captures/stopped-working-tree-inventories/delete:
    post:
      consumes:
        - application/json
      operationId: DeleteWorkingTreeInventory
      parameters:
        - description: Sandbox ID
          in: path
          name: sandboxId
          required: true
          type: string
        - description: Exact stopped working-tree inventory authority
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/workingcopy.WorkingTreeInventoryRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/workingcopy.WorkingTreeInventoryDeletionReceipt'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '503':
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - Bearer: []
      summary: Delete exact inventory custody and prove its absence
      tags:
        - sandbox
...
security:
  - Bearer: []
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and an API token.
    in: header
    name: Authorization
    type: apiKey
swagger: '2.0'